Made unconditional deletion on copyfiles into conditional --
git-svn-id: https://cida-svn.er.usgs.gov/repos/dev/usgs/gdp/trunk@9172 622dc1fb-2224-0410-ab72-dc85a3975e1a

+++ src/main/java/gov/usgs/gdp/helper/FileHelper.java
@@ -42,7 +42,7 @@ public class FileHelper {
 		} else {
 			FileUtils.copyFile(inFile, (new File(outPath + FileHelper.getSeparator() + inFile.getName())));
 		}
-		FileUtils.deleteQuietly(inFile);
+		if (deleteOriginalFile) FileUtils.deleteQuietly(inFile);
 		return true;
 	}
